==> American Express - Default Prediction/7. Combine.ipynb <==
# 7. Combine

import numpy as np
import pandas as pd

## Load Predictions

submit = pd.read_csv('../input/amex-default-prediction/sample_submission.csv')
submit['prediction'] = 0

xgb0 = pd.read_csv('../sub/submission_xgb.csv') 
lgb1 = pd.read_csv('../sub/test_lgbm_5fold_seed42.csv') 
lgb2 = pd.read_csv('../sub/test_lgbm_5fold_seed42_v2.csv') 
lgb3 = pd.read_csv('../sub/test_lgbm_5fold_seed42_v3.csv') 

## Blending: Linear Weighted Sum

submit['prediction'] =  xgb0['prediction']*0.2 + lgb1['prediction']*0.25 +lgb2['prediction']*0.25 +lgb3['prediction']*0.3
submit.to_csv('submission.csv', index=None)
